[refactoring] rearrange components file structure.
