+++ frontend/app/components/resource/resource.scss
