+++ frontend/app/components/timeline/timeline.scss
@@ -36,7 +36,7 @@
   top: 60px;
 }
 
-#timeline-header {
+.timeline-header {
   position: absolute;
   top: 0;
   left: 0;
@@ -50,7 +50,7 @@
   background-repeat: no-repeat;
 }
 
-#timeline-scales {
+.timeline-scales {
   position: absolute;
   width: 100%;
 }
