+++ frontend/package.json
@@ -27,8 +27,11 @@
     "gulp-sourcemaps": "^1.5.0",
     "gulp-uglify": "^1.1.0",
     "gulp-useref": "^1.1.1",
+    "html-loader": "^0.3.0",
     "main-bower-files": "^2.5.0",
+    "node-sass": "^3.2.0",
     "opn": "^1.0.1",
+    "sass-loader": "^1.0.2",
     "webpack-stream": "^2.0.0",
     "wiredep": "^2.2.2"
   },
